Keep both tabs in memory separately
This allows the user to swipe back and forth between Top and New. It also allows easy adding of additional tabs, if desired (this could be a setting, not implemented yet).

The ChangeNotifiers are now more granular. `HackerNewsTab` is a changenotifier that lives under `HackerNewsNotifier`. We also have a very simple `LoadingTabsCount` value notifier, for showing when any of the tabs (or both) are currently loading. This is used in the `LoadingInfo` widget.

The `LoadingInfo` widget shows a bit more visible animation — the previous “pulse” was a bit too subtle, to the point of not being seen by most users at all.