Move fetch and parse to isolate
Episode 32 code, finishing work from episode 31. Fetching of article ids and articles themselves, as well as parsing, is not happening on the main (UI) thread anymore.

Removed the caching of articles (might reintroduce this again, but the problem was too gnarly for the allotted time).

The Worker code is not properly tested. We debated if injecting an HttpClient would work across isolates.