Fix bug with unordered particles
Articles were coming back in parallel, and added to the `results` as they came in. Which meant that, often, the sorting of the articles in the app would not match the actual sorting of the articles from the API.

The fix is to resort the articles by comparing their ID’s position in the initial list of ids to each other.